Rework and simplify Sandblast
As a consequence of removing the player's ability to wield non-weapons,
the player can no longer wield stones or large rocks... which means the
three different use-cases for Sandblast needs to be condensed into one.
(Frankly, this is probably the right move anyway: consistency in spell
behavior is good.)

The new Sandblast works like Sticks to Snakes: consumes a missile in
your inventory to power the spell. There is no fallback anymore for this
spell, if you're out of stones you can no longer cast Sandblast.